Safari should be killed at end of test run
Calling "open" to launch safari works well to open Safari, but does not allow
Safari to be killed by Karma when it is finished with tests. We can call Safari
directly (and therefore be able to kill the process) by using a local temp file
for redirection to the provided url.

Change-Id: I5f1d5b9e6418d1855ca72d5b9e68694fd8abc377